+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/life/ActivityTemplateDriverInfoController.java
@@ -1,0 +1,148 @@
+package com.dwarfeng.familyhelper.webapi.node.controller.v1.life;
+
+import com.dwarfeng.familyhelper.life.sdk.bean.entity.JSFixedFastJsonActivityTemplateDriverInfo;
+import com.dwarfeng.familyhelper.life.sdk.bean.entity.WebInputActivityTemplateDriverInfo;
+import com.dwarfeng.familyhelper.life.stack.bean.entity.ActivityTemplateDriverInfo;
+import com.dwarfeng.familyhelper.webapi.stack.service.life.ActivityTemplateDriverInfoResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.JSFixedFastJsonPagedData;
+import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.bean.key.FastJsonLongIdKey;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.sdk.interceptor.http.BindingCheck;
+import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
+import com.dwarfeng.subgrade.sdk.validation.group.Insert;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * 提醒驱动器信息控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.10
+ */
+@RestController
+@RequestMapping("/api/v1/life")
+public class ActivityTemplateDriverInfoController {
+
+    private final ActivityTemplateDriverInfoResponseService service;
+    private final ServiceExceptionMapper sem;
+
+    private final BeanTransformer<ActivityTemplateDriverInfo, JSFixedFastJsonActivityTemplateDriverInfo>
+            beanTransformer;
+
+    public ActivityTemplateDriverInfoController(
+            ActivityTemplateDriverInfoResponseService service, ServiceExceptionMapper sem,
+            BeanTransformer<ActivityTemplateDriverInfo, JSFixedFastJsonActivityTemplateDriverInfo> beanTransformer
+    ) {
+        this.service = service;
+        this.sem = sem;
+        this.beanTransformer = beanTransformer;
+    }
+
+    @GetMapping("/activity-template-driver-info/{id}/exists")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<Boolean> exists(HttpServletRequest request, @PathVariable("id") long id) {
+        try {
+            boolean exists = service.exists(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/activity-template-driver-info/{id}")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonActivityTemplateDriverInfo> get(
+            HttpServletRequest request, @PathVariable("id") long id
+    ) {
+        try {
+            ActivityTemplateDriverInfo activityTemplateDriverInfo = service.get(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(
+                    JSFixedFastJsonActivityTemplateDriverInfo.of(activityTemplateDriverInfo)
+            ));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @PostMapping("/activity-template-driver-info")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public FastJsonResponseData<FastJsonLongIdKey> insert(
+            HttpServletRequest request,
+            @RequestBody @Validated(Insert.class) WebInputActivityTemplateDriverInfo webInputActivityTemplateDriverInfo,
+            BindingResult bindingResult
+    ) {
+        try {
+            ActivityTemplateDriverInfo activityTemplateDriverInfo =
+                    WebInputActivityTemplateDriverInfo.toStackBean(webInputActivityTemplateDriverInfo);
+            LongIdKey insert = service.insert(activityTemplateDriverInfo);
+            return FastJsonResponseData.of(ResponseDataUtil.good(FastJsonLongIdKey.of(insert)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @PatchMapping("/activity-template-driver-info")
+    @BehaviorAnalyse
+    @BindingCheck
+    @LoginRequired
+    public FastJsonResponseData<Object> update(
+            HttpServletRequest request,
+            @RequestBody @Validated WebInputActivityTemplateDriverInfo webInputActivityTemplateDriverInfo,
+            BindingResult bindingResult
+    ) {
+        try {
+            service.update(WebInputActivityTemplateDriverInfo.toStackBean(webInputActivityTemplateDriverInfo));
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @DeleteMapping("/activity-template-driver-info/{id}")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<Object> delete(HttpServletRequest request, @PathVariable("id") long id) {
+        try {
+            service.delete(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/activity-template/{activityTemplateId}/activity-template-driver-info")
+    @BehaviorAnalyse
+    @SkipRecord
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonActivityTemplateDriverInfo>>
+    childForActivityTemplate(
+            HttpServletRequest request, @PathVariable("activityTemplateId") Long activityTemplateId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<ActivityTemplateDriverInfo> all = service.childForActivityTemplate(
+                    new LongIdKey(activityTemplateId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonActivityTemplateDriverInfo> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+}