+++ CHANGELOG.md
@@ -28,6 +28,7 @@
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTemplateParticipantController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTemplateFileController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTemplateDataInfoController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTemplateDriverInfoController。
 
 - 为 dubbo 的消费者增加了分组设置。
   - snowflake。